9.30 commit every thing from begin

+++ game/src/main/java/net/demilich/metastone/game/behaviour/LinearBatchSA.java
@@ -104,7 +104,7 @@ public class LinearBatchSA extends Behaviour{
         if (opponent.getHero().isDestroyed()) {  // 对方被干掉，得分 正无穷
             return Float.POSITIVE_INFINITY;
         }
-        List<Integer> envState = player.getPlayerStatefh0(false);
+        List<Double> envState = player.getPlayerStatefh0(false);
         envState.addAll(opponent.getPlayerStatefh0(false));
 
         double score = 0;
