+++ game/src/main/java/net/demilich/metastone/game/behaviour/LinearCEM.java
@@ -100,7 +100,7 @@ public class LinearCEM extends Behaviour {
 		if (opponent.getHero().isDestroyed()) {  // 对方被干掉，得分 正无穷
 			return Float.POSITIVE_INFINITY;
 		}
-		List<Integer> envState = player.getPlayerStatefh0(false);
+		List<Double> envState = player.getPlayerStatefh0(false);
 		envState.addAll(opponent.getPlayerStatefh0(false));
 //		logger.info("Par Weight: {}, envState: {}", parWeight.length, envState.size());
 		double score = 0;
