+++ game/src/main/java/net/demilich/metastone/game/behaviour/LinearES.java
@@ -120,7 +120,7 @@ public class LinearES extends Behaviour {
         if (opponent.getHero().isDestroyed()) {  // 对方被干掉，得分 正无穷
             return Float.POSITIVE_INFINITY;
         }
-        List<Integer> envState = player.getPlayerStatefh0(false);
+        List<Double> envState = player.getPlayerStatefh0(false);
         envState.addAll(opponent.getPlayerStatefh0(false));
 
         assert (envState.size() == feaNum);
